Split runner agent sink tests
